requirements.txt
gunicorn==19.7.1
oauth2client==4.0.0
newspaper3k==0.2.2
Flask==0.12.2
requests==2.18.1
google-cloud-datastore==1.1.0
google-cloud==0.26.1
numpy==1.13.1
scipy==0.19.1
scikit-learn==0.18.2
beautifulsoup4==4.6.0
tweepy==3.5.0
nltk==3.2.4
regex==2017.7.11
